Rename `src/util/message.cpp` to `src/util/strings.cpp` and add appropriate headers (#32379)
* Move things from message, update dependencies

* Fix lint

* Restyle

* Remove wrongly copied comment

* Add another dependency

* Vendor identifiers is not orphaned anymore

* Move strings to ember-strings to not have overlap in names

* Add include for the cpp file into its own header

* Standard ordering for attribute metadata cpp

* Update darwin framework xcode file

* Also move the copy string functions into ember strings. header af.h and implementation in util.cpp makes no sense
